refactor: QuizCoachmark > 상수 파일로 분리
[#304]
