+++ packages/frontend/src/components/quiz/QuizCoachmark/coachmarkData.ts
@@ -1,0 +1,6 @@
+export const COACHMARK_TARGETS = {
+  GIT_GRAPH: "coach--git-graph",
+  KEY_COMMAND: "coach--key-command",
+  TERMINAL: "coach--terminal",
+  RESIZABLE: "coach--resizable",
+};
